chore: Move patterns logic to top-level file.
