+++ ee/hogai/session_summaries/patterns.py
